[REFACTOR] modularize backend routes
